refactor: change interface filename to types
