+++ apps/design-system/src/components/Table/types.ts
