+++ libs/core-ui/src/components/AnnouncementBar/types.ts
