+++ libs/core-ui/src/components/Hero/types.ts
