+++ libs/core-ui/src/components/Typography/index.ts
@@ -1,4 +1,4 @@
 export * from './Headings';
 export * from './Paragraph';
 export * from './Eyebrow';
-export * from './interface';
+export * from './types';
