+++ libs/core-ui/src/index.ts
@@ -1,5 +1,4 @@
 export * from './components/Button';
-export * from './components/Text';
 export * from './components/AnnouncementBar';
 export * from './components/Header';
 export * from './components/Header';
