+++ tools/generators/src/generators/component/files/types.template
